==> 1-exploration.ipynb <==
# project part 1 - data exploration
# sections:
# 1. data profiling
# 2. data cleaning
# 3. data visualization

***

# import libraries

# for data manipulation
import math
import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np
import datetime as dt
from scipy import stats as st

# for data visualization
import matplotlib.pyplot as plt
import seaborn as sns

# import data
df = pd.read_excel("data/credit card_defaults.xlsx")

***

# data profiling

# preview data - head
df.head()

# preview data - tail
df.tail()

# data shape
df.shape

# data info
df.info()

# data describe
df.describe(include='all').T

***

# data cleaning

# check missing values
df.isna().sum()

# check duplicate values
df.duplicated().sum()

# rename `default payment this month` to `RESPONSE`
df = df.rename(columns={'default payment this month':'RESPONSE'})

# check changes made
df.columns

# squeeze 'others' into 4/5/6 to reduce variance in EDUCATION column - reduce underfitting
df['EDUCATION'] = df['EDUCATION'].replace([0, 4, 5, 6], 4)

# create new column for visualization 
df['AGEBIN'] = pd.cut(df['AGE'],[20,25, 30, 35, 40, 50, 60, 80])
df['AGEBIN'] = df['AGEBIN'].astype('str')
df['AGEBIN'] = df['AGEBIN'].replace({'(20, 25]':'21-25',
                                     '(25, 30]':'26-30',
                                     '(30, 35]':'31-35',
                                     '(35, 40]':'36-40',
                                     '(40, 50]':'41-50',
                                     '(50, 60]':'51-60',
                                     '(60, 80]':'61-80'
                                    })
agebin_order = ['21-25', '26-30', '31-35', '36-40', '41-50', '51-60', '61-80']

# check for inactive customers
# we shouldn't use this data because there's nothing to predict on
inactive = df[(df['BILL_AMT1']==0) & (df['BILL_AMT2']==0) & (df['BILL_AMT3']==0) & 
              (df['BILL_AMT4']==0) & (df['BILL_AMT5']==0) &(df['BILL_AMT6']==0)]
inactive.shape
# drop index of inactive customers
df = df.drop(inactive.index).reset_index().drop(columns='index')
df

***

# data exploration - categorical variables

# describe data - RESPONSE
df['RESPONSE'].value_counts(normalize=True)

# describe data - SEX
df['SEX'].value_counts()

# barplot - SEX
plt.figure(figsize=(15,4))

ax = sns.countplot(data = df, x = 'SEX', hue="RESPONSE")

plt.xlabel("Sex")
plt.ylabel("# of Clients")
plt.ylim(0,20000)
plt.xticks([0,1],['Male', 'Female'])
plt.show()

# describe data - EDUCATION
df['EDUCATION'].value_counts()

# barplot - EDUCATION
plt.figure(figsize=(15,4))
ax = sns.countplot(data=df, x='EDUCATION', hue="RESPONSE")

plt.xlabel("Education")
plt.ylabel("# of Clients")
plt.ylim(0,12000)
plt.xticks([0,1,2,3],['Grad School','University','High School','Others'])
plt.show()

# describe data - MARRIAGE
df['MARRIAGE'].value_counts()

# barplot - MARRIAGE
plt.figure(figsize=(15,4))
ax = sns.countplot(data=df, x='MARRIAGE', hue="RESPONSE")

plt.xlabel("Marital Status")
plt.ylabel("# of Clients")
plt.ylim(0,15000)
plt.xticks([0,1,2,3],['Unknown', 'Married', 'Single', 'Divorce'])
plt.show()

***

# data exploration - numerical variables

# describe data - AGE
df[['AGE']].describe().T

# barplot - AGE
plt.figure(figsize=(15,4))
ax = sns.countplot(data=df, x='AGEBIN', hue="RESPONSE", order=agebin_order)

plt.xlabel("Age Group")
plt.ylabel("# of Clients")
plt.ylim(0,8000)
plt.show()

# describe data - LIMIT BAL
df[['LIMIT_BAL']].describe().T

# boxplot - LIMIT BAL
plt.figure(figsize=(17,2))
plt.title('Bill Amount')
sns.boxplot(x=df['LIMIT_BAL'])

# kde dsitribution plot - LIMIT BAL
plt.figure(figsize=(15,4))
sns.kdeplot(df.loc[(df['RESPONSE']==0), 'LIMIT_BAL'], label='No Default', shade=True)
sns.kdeplot(df.loc[(df['RESPONSE']==1), 'LIMIT_BAL'], label='Default', shade=True)
plt.ticklabel_format(style='plain', axis='x') #repressing scientific notation on x
plt.ylabel('Density')
plt.legend()
plt.show()

# describe data - BILL AMT
df[['BILL_AMT1', 'BILL_AMT2', 'BILL_AMT3', 'BILL_AMT4', 'BILL_AMT5', 'BILL_AMT6']].describe().T

# checking for outliers - BILL AMT
df[(df["BILL_AMT1"]<0) | (df["BILL_AMT2"]<0) | (df["BILL_AMT3"]<0) | (df["BILL_AMT4"]<0) | (df["BILL_AMT5"]<0) | (df["BILL_AMT6"]<0)].shape[0]/30000 * 100

# boxplot - BILL AMT
# melt transform from wide to long
billamt_df = pd.melt(df, id_vars=['RESPONSE'], 
                         value_vars=['BILL_AMT1', 'BILL_AMT2', 'BILL_AMT3', 'BILL_AMT4', 'BILL_AMT5', 'BILL_AMT6'])

# plot
plt.figure(figsize=(17,9))
plt.title('Bill Amount')
sns.boxplot(data=billamt_df, x="value", y="variable", hue="RESPONSE", gap=.1)

# null hypothesis: the mean BILL_AMTn for defaulters and non-defaulters are the same
# alternative Hypothesis: the mean BILL_AMTn for defaulters and non-defaulters are different

cols = ['BILL_AMT1', 'BILL_AMT2', 'BILL_AMT3', 'BILL_AMT4', 'BILL_AMT5', 'BILL_AMT6']

for col in cols:
    # create data sample
    sample_0 = np.asarray(df[[col, 'RESPONSE']].query('RESPONSE == 0')[col].to_list())
    sample_1 = np.asarray(df[[col, 'RESPONSE']].query('RESPONSE == 1')[col].to_list())

    # define significance level
    alpha = 0.05 

    # test for equality of variances
    lv = st.levene(sample_0, sample_1)
    if lv.pvalue > alpha: param = True
    else: param = False

    # t-test for differences in sample mean
    results = st.ttest_ind(sample_0, sample_1, equal_var=param)

    # print results
    print('p-value:', results.pvalue)
    if results.pvalue < alpha: print(col, " Reject the null hypothesis")
    else: print(col, " Failed to reject null hypothesis")

# describe data - PAY AMT
df[['PAY_AMT1', 'PAY_AMT2', 'PAY_AMT3', 'PAY_AMT4', 'PAY_AMT5', 'PAY_AMT6']].describe().T

# box and whiskers plot - PAY AMT
# melt to transform from wide to long
payamt_df = pd.melt(df, id_vars=['RESPONSE'],
                        value_vars=['PAY_AMT1', 'PAY_AMT2', 'PAY_AMT3', 'PAY_AMT4', 'PAY_AMT5', 'PAY_AMT6'])

# boxplot
plt.figure(figsize=(17,9))
plt.title('Pay Amount')
sns.boxplot(data=payamt_df, 
            x="value", 
            y="variable", 
            hue="RESPONSE")

# null hypothesis: the mean PAY_AMTn for defaulters and non-defaulters are the same
# alternative Hypothesis: the mean PAY_AMTn for defaulters and non-defaulters are different

cols = ['PAY_AMT1', 'PAY_AMT2', 'PAY_AMT3', 'PAY_AMT4', 'PAY_AMT5', 'PAY_AMT6']

for col in cols:
    # create data sample
    sample_0 = np.asarray(df[[col, 'RESPONSE']].query('RESPONSE == 0')[col].to_list())
    sample_1 = np.asarray(df[[col, 'RESPONSE']].query('RESPONSE == 1')[col].to_list())

    # define significance level
    alpha = 0.05 

    # test for equality of variances
    lv = st.levene(sample_0, sample_1)
    if lv.pvalue > alpha: param = True
    else: param = False

    # t-test for differences in sample mean
    results = st.ttest_ind(sample_0, sample_1, equal_var=param)

    # print results
    print('p-value:', results.pvalue)
    if results.pvalue < alpha: print(col, " Reject the null hypothesis")
    else: print(col, " Failed to reject null hypothesis")

***

# export cleaned and processed data 
df.to_excel("/data/data_cleaned.xlsx")
