==> 3-clustering.ipynb <==
# part 3 - clustering

# import data
df = pd.read_excel()

***

X = df4[df4['RESPONSE']==1][['LIMIT_BAL', 'AGE', 'SEX', 'EDUCATION', 'MARRIAGE', 'BILL_AMT_TOTAL', 'PAY_AMT_TOTAL',
                             'OUTS_AMT_TOTAL', 'PAST_DEFAULTS', 'EXCEED_LIMIT_BAL']]
X_clustered = X.copy()
X.head(1)

# data scaling
num_col = X.select_dtypes(include=np.number).columns.tolist()

# scaler
scaler = MinMaxScaler()
X[num_col] = scaler.fit_transform(X[num_col])

# finding best k
cost = []
sil = []

k_range = range(2,10)
for k in list(k_range):
    kmodes = KModes(n_clusters=k, init="random", n_init=5, random_state=42)
    kmodes.fit_predict(X)
    cost.append(kmodes.cost_)
    sil.append(silhouette_score(X, kmodes.labels_))

# plot
plt.plot(k_range, cost, 'bx-')
plt.xlabel('No. of clusters')
plt.ylabel('Cost')
plt.title('Elbow Method For Optimal k')
plt.show()

# silhouette plot
plt.plot(k_range, sil, 'bx-')
plt.xlabel('No. of clusters')
plt.ylabel('Cost')
plt.title('Silhouette score against K')
plt.show()

***

# build a model with 4 clusters
kmode = KModes(n_clusters=4, init='random', n_init=5, verbose=1)
clusters = kmode.fit_predict(X)
clusters

X_clustered['CLUSTER'] = clusters
X_clustered

X_clustered['CLUSTER'].value_counts()

# we have to map the indexes back to the original unscaled dataframe to do some data exploration
cluster1_index = X_clustered[X_clustered['CLUSTER']==0].index
cluster2_index = X_clustered[X_clustered['CLUSTER']==1].index
cluster3_index = X_clustered[X_clustered['CLUSTER']==2].index
cluster4_index = X_clustered[X_clustered['CLUSTER']==3].index

cluster1 = X_clustered[X_clustered.index.isin(cluster1_index)]
cluster2 = X_clustered[X_clustered.index.isin(cluster2_index)]
cluster3 = X_clustered[X_clustered.index.isin(cluster3_index)]
cluster4 = X_clustered[X_clustered.index.isin(cluster4_index)]

cluster1.head()

# cluster 1 - descriptive statistics
round(cluster1.describe().T,2)

limit_bal_med = [cluster1['LIMIT_BAL'].median(), cluster2['LIMIT_BAL'].median(), cluster3['LIMIT_BAL'].median(), cluster4['LIMIT_BAL'].median()]
age_med = [cluster1['AGE'].median(), cluster2['AGE'].median(), cluster3['AGE'].median(), cluster4['AGE'].median()]
bill_amt_total_med = [cluster1['BILL_AMT_TOTAL'].median(), cluster2['BILL_AMT_TOTAL'].median(), cluster3['BILL_AMT_TOTAL'].median(), cluster4['BILL_AMT_TOTAL'].median()]
pay_amt_total_med = [cluster1['PAY_AMT_TOTAL'].median(), cluster2['PAY_AMT_TOTAL'].median(), cluster3['PAY_AMT_TOTAL'].median(), cluster4['PAY_AMT_TOTAL'].median()]
outs_amt_total_med = [cluster1['OUTS_AMT_TOTAL'].median(), cluster2['OUTS_AMT_TOTAL'].median(), cluster3['OUTS_AMT_TOTAL'].median(), cluster4['OUTS_AMT_TOTAL'].median()]
past_defaults_med = [cluster1['PAST_DEFAULTS'].median(), cluster2['PAST_DEFAULTS'].median(), cluster3['PAST_DEFAULTS'].median(), cluster4['PAST_DEFAULTS'].median()]
exceed_limit_bal_med = [cluster1['EXCEED_LIMIT_BAL'].median(), cluster2['EXCEED_LIMIT_BAL'].median(), cluster3['EXCEED_LIMIT_BAL'].median(), cluster4['EXCEED_LIMIT_BAL'].median()]
sex_mode = [cluster1['SEX'].mode().tolist()[0], cluster2['SEX'].mode().tolist()[0], cluster3['SEX'].mode().tolist()[0], cluster4['SEX'].mode().tolist()[0]]
education_mode = [cluster1['EDUCATION'].mode().tolist()[0], cluster2['EDUCATION'].mode().tolist()[0], cluster3['EDUCATION'].mode().tolist()[0], cluster4['EDUCATION'].mode().tolist()[0]]
marriage_mode = [cluster1['MARRIAGE'].mode().tolist()[0], cluster2['MARRIAGE'].mode().tolist()[0], cluster3['MARRIAGE'].mode().tolist()[0], cluster4['MARRIAGE'].mode().tolist()[0]]

cluster_characteristics = pd.DataFrame(data = {"CLUSTER" : range(1,5),
                                               "LIMIT_BAL" : limit_bal_med,
                                               "AGE"       : age_med,
                                               "BILL_AMT_TOTAL" : bill_amt_total_med,
                                               "PAY_AMT_TOTAL" : pay_amt_total_med,
                                               "OUTS_AMT_TOTAL": outs_amt_total_med,
                                               "PAST_DEFAULTS" : past_defaults_med,
                                               "EXCEED_LIMIT_BAL" : exceed_limit_bal_med,
                                               "SEX" : sex_mode,
                                               "EDUCATION" : education_mode,
                                               "MARRIAGE" : marriage_mode})
cluster_characteristics
